added logout button and made quarter hours work

+++ server/methods.js
@@ -273,7 +273,7 @@ Meteor.startup(function () {
                                                 var entry = sheet.projectEntriesArray[pIndex].EntryArray[eIndex];
                                                 var sum = 0;
                                                 for (var hours in entry.hours) {
-                                                    sum += parseInt(entry.hours[hours]);
+                                                    sum += parseFloat(entry.hours[hours]);
                                                 }
                                                 projectHours[currentProject] += sum;
                                                 projectComments[currentProject].push(entry.Comment);
